requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.16
django-colorfield==0.7.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.38.0
html5lib==1.1
idna==3.4
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pyphen==0.13.1
python3-openid==3.2.0
pytz==2022.5
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tinycss2==1.2.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
weasyprint==57.1
webcolors==1.12
webencodings==0.5.1
zipp==3.10.0
zopfli==0.2.2
